# tsm1 index fixes and improvements