--- a/tsdb/tsm1/file_store_key_iterator.go
+++ b/tsdb/tsm1/file_store_key_iterator.go
@@ -25,10 +25,13 @@ func (k *keyIterator) next() bool {
 	return false
 }
 
+func (k *keyIterator) Err() error { return k.iter.Err() }
+
 type mergeKeyIterator struct {
 	itrs keyIterators
 	key  []byte
 	typ  byte
+	err  error
 }
 
 func newMergeKeyIterator(files []TSMFile, seek []byte) *mergeKeyIterator {
@@ -46,6 +49,10 @@ func newMergeKeyIterator(files []TSMFile, seek []byte) *mergeKeyIterator {
 }
 
 func (m *mergeKeyIterator) Next() bool {
+	if m.err != nil {
+		return false
+	}
+
 	merging := len(m.itrs) > 1
 
 RETRY:
@@ -55,6 +62,13 @@ RETRY:
 
 	key, typ := m.itrs[0].key, m.itrs[0].typ
 	more := m.itrs[0].next()
+
+	if !more {
+		if err := m.itrs[0].Err(); err != nil {
+			m.err = err
+			return false
+		}
+	}
 
 	switch {
 	case len(m.itrs) > 1:
@@ -77,9 +91,10 @@ RETRY:
 	}
 
 	m.key, m.typ = key, typ
-
 	return true
 }
+
+func (m *mergeKeyIterator) Err() error { return m.err }
 
 func (m *mergeKeyIterator) Read() ([]byte, byte) { return m.key, m.typ }