--- a/tsdb/tsm1/file_store_key_iterator.go
+++ b/tsdb/tsm1/file_store_key_iterator.go
@@ -6,33 +6,20 @@ import (
 )
 
 type keyIterator struct {
-	f   TSMFile
-	c   int // current key index
-	n   int // key count
-	key []byte
-	typ byte
+	iter TSMIterator
+	key  []byte
+	typ  byte
 }
 
 func newKeyIterator(f TSMFile, seek []byte) *keyIterator {
-	c, n := 0, f.KeyCount()
-	if len(seek) > 0 {
-		c = f.Seek(seek)
-	}
-
-	if c >= n {
-		return nil
-	}
-
-	k := &keyIterator{f: f, c: c, n: n}
+	k := &keyIterator{iter: f.Iterator(seek)}
 	k.next()
-
 	return k
 }
 
 func (k *keyIterator) next() bool {
-	if k.c < k.n {
-		k.key, k.typ = k.f.KeyAt(k.c)
-		k.c++
+	if k.iter.Next() {
+		k.key, k.typ = k.iter.Key(), k.iter.Type()
 		return true
 	}
 	return false
@@ -98,9 +85,10 @@ func (m *mergeKeyIterator) Read() ([]byte, byte) { return m.key, m.typ }
 
 type keyIterators []*keyIterator
 
-func (k keyIterators) Len() int            { return len(k) }
-func (k keyIterators) Less(i, j int) bool  { return bytes.Compare(k[i].key, k[j].key) == -1 }
-func (k keyIterators) Swap(i, j int)       { k[i], k[j] = k[j], k[i] }
+func (k keyIterators) Len() int           { return len(k) }
+func (k keyIterators) Less(i, j int) bool { return bytes.Compare(k[i].key, k[j].key) == -1 }
+func (k keyIterators) Swap(i, j int)      { k[i], k[j] = k[j], k[i] }
+
 func (k *keyIterators) Push(x interface{}) { *k = append(*k, x.(*keyIterator)) }
 
 func (k *keyIterators) Pop() interface{} {