# Summary

TL;DR:

- Fixes #1983 
- Improves benchmarks and reduces page faults
- Each commit is self contained for review if desired

---

### Remove offsets from mmap
This uses a `[]uint32` instead of a `[]byte` stored off the Go heap.

### Use a uint32 for offsets globally
Consolidates offsets from `int`, `int32`, `uint32` into just `uint32`.

### Use uint32 key for tombstones
Rather than look up tombstones by `[]byte`, look them up by the `uint32` offset that almost certainly have access to already.

### Fix multiple issues with DeleteRange
DeleteRange wasn't precise about keys being deleted, had incorrect locking, and didn't seek when the next key to delete wasn't the next key in the index.

### Add helper to track page faults in index
Estimates page faults by keeping track of the most recently accessed page, and considers it a fault if the accessed page is not the most recently accessed, or the one after it.

### Keep first 8 bytes of each key in memory
Reduces the amount of page faults by ~30% at very little memory overhead. Since keys tend to share the same first 8 bytes, rather than store the 8 bytes for every key, we do a sort of run length encoding, so we only store 8 bytes for every distinct prefix.

### Introduce readerOffsets to manage the offsets slice
Just an abstraction around the offsets and prefixes slice to keep the delete logic in one spot and well tested.

### Use readerOffsetsIterator for deletes
Cleans up a bunch of code in the Delete paths because of the previous abstraction.

### Change TSMFile to use an iterator style api
Direct indexes become invalidated during deletes. This returns an iterator so that we can add code to make it safe to iterate while deletes are in progress.

# Benchmarks

```
name                           old time/op    new time/op    delta
UnmarshalBinary                  70.0ms ±17%    71.0ms ±12%      ~     (p=1.000 n=8+8)
DeleteRangeLast                  1.48µs ± 1%    0.28µs ± 5%   -81.29%  (p=0.000 n=8+7)
DeleteRangeFull/Large             786ms ± 1%     363ms ± 3%   -53.89%  (p=0.000 n=7+8)
DeleteRangeFull/Small            2.37ms ± 0%    1.14ms ± 3%   -52.02%  (p=0.000 n=7+8)
DeleteRangeFull_Covered/Large     384ms ± 2%     188ms ± 3%   -51.04%  (p=0.000 n=8+8)
DeleteRangeFull_Covered/Small     470µs ± 1%     190µs ± 1%   -59.71%  (p=0.000 n=8+7)
Delete/Large                     74.0ms ± 1%   128.7ms ± 1%   +73.80%  (p=0.001 n=7+7)
Delete/Small                      142µs ± 1%     130µs ± 1%    -8.24%  (p=0.000 n=8+8)

name                           old alloc/op   new alloc/op   delta
UnmarshalBinary                  11.6MB ± 0%    11.7MB ± 0%    +0.02%  (p=0.000 n=8+7)
DeleteRangeLast                  3.26kB ± 0%   0.00kB ±NaN%  -100.00%  (p=0.000 n=8+8)
DeleteRangeFull/Large             233MB ± 0%     161MB ± 0%   -30.75%  (p=0.000 n=8+8)
DeleteRangeFull/Small            2.13MB ± 0%    1.40MB ± 0%   -34.53%  (p=0.000 n=8+8)
DeleteRangeFull_Covered/Large    12.4MB ± 0%     0.4MB ± 0%   -96.82%  (p=0.002 n=7+8)
DeleteRangeFull_Covered/Small     120kB ± 0%       0kB ± 0%   -99.89%  (p=0.000 n=8+8)
Delete/Large                     4.54kB ± 0%    0.21kB ± 0%   -95.26%  (p=0.000 n=8+8)
Delete/Small                      80.0B ± 0%     0.0B ±NaN%  -100.00%  (p=0.000 n=8+8)

name                           old allocs/op  new allocs/op  delta
UnmarshalBinary                    35.0 ± 0%      42.0 ± 0%   +20.00%  (p=0.000 n=8+8)
DeleteRangeLast                    3.00 ± 0%     0.00 ±NaN%  -100.00%  (p=0.000 n=8+8)
DeleteRangeFull/Large             1.53M ± 0%     0.52M ± 0%   -65.98%  (p=0.000 n=8+8)
DeleteRangeFull/Small             15.2k ± 0%      5.2k ± 0%   -65.97%  (p=0.000 n=8+8)
DeleteRangeFull_Covered/Large       620 ± 0%       124 ± 0%   -80.00%  (p=0.002 n=7+8)
DeleteRangeFull_Covered/Small      10.0 ± 0%       2.0 ± 0%   -80.00%  (p=0.000 n=8+8)
Delete/Large                        246 ± 0%         1 ± 0%   -99.59%  (p=0.000 n=8+8)
Delete/Small                       4.00 ± 0%     0.00 ±NaN%  -100.00%  (p=0.000 n=8+8)
```

The below numbers measure an estimate of page faults, so lower speeds are better. It
is considered a fault if the access is not the most recently access page, or the next
page after the most recently accessed page.

```
DeleteRangeFull          24.7MB/s ±10%  17.4MB/s ± 7%   -29.54%          (p=0.000 n=8+8)
DeleteRangeFull_Covered  14.2MB/s ± 1%     0MB/s ± 0%  -100.00%          (p=0.000 n=8+7)
Delete                    171MB/s ± 1%     0MB/s ±14%   -99.97%          (p=0.000 n=7+8)
```

DeleteRangeFull_Covered is the case where it deletes every key but without that being
statically known: it's just a range large enough to cover every entry. While it gets
slower in terms of cpu time, it's deleting 500k series, so it's really about 1.4µs per
series. The wins from being more linear in disk access seem more important.